refactor(core): fix layer violation — extract ChatEvent to core/types/chat_event.ts
ChatEvent, ChatClientMessage, and ChatEventSender were defined in
gateway/chat_types.ts but imported by channels/cli/ files, violating the
rule that channels/ should not import from gateway/.

Extracted types to core/types/chat_event.ts. Gateway module re-exports
for backward compatibility.
